requirements.txt
PySide6>=6.0.0
numpy>=1.21.0
PyOpenGL>=3.1.5
pyrr>=0.9.5